Apply variant masks to annotation file

This function applies variant mask criteria to a data frame containing variant annotations. The function can accept multiple sets of filter criteria, each defined as a list of column names and filter conditions. The resulting filtered data frames are returned as a named list.

apply_variant_masks(annotation_df, masks)

Arguments

annotation_df

A data frame to be filtered.

masks

A named list of lists containing filter criteria. Each element of the outer list should be named after the desired output name for the filtered data frame (eg. name of the variant mask). Each inner list should contain the filter criteria for a specific set of columns. For example: list(plof_0.001 = list(ExonicFunc.ensGene = "= stopgain", gnomAD_exome_ALL = "< 0.001"), common_0.01 = list(gnomAD_exome_ALL = "> 0.01"))

Value

A named list of filtered tibble::tibble()s based on the provided filter criteria.

Examples

if (FALSE) {

# Extract variants in a gene
ldlr_variants <- awk_str_filter(
  filename = "/project/PMBB/PMBB-Release-2020-2.0/Exome/Variant_annotations/PMBB-Release-2020-2.0_genetic_exome_variant-annotation-counts.txt",
  filter_col = "Gene.refGene",
  filter_str = c("LDLR"))

# Apply the specified masks to the set of variants
apply_variant_masks(
  ldlr_variants,
  masks = list(
    plof_0.001 = list(ExonicFunc.ensGene = "== 'stopgain'", gnomAD_exome_ALL = "< 0.001"),
    common_0.01 = list(gnomAD_exome_ALL = "> 0.01"),
    exonic_0.001 = list(Func.refGene = "== 'exonic'", gnomAD_exome_ALL = "< 0.001")
  )
)
}