Extract genotypes from PMBB plink files

This function is a wrapper around plink2 to extract genotypes from PMBB plink files. The function accepts the path to the plink files and the list of variant IDs to extract. The function returns a long-format tibble::tibble() containing the extracted genotypes for each PMBB participant.

pmbb_extract_genotypes(
  variant_df,
  variant_id_col,
  effect_allele_col,
  plink_bin,
  bfile,
  threads = 1,
  memory = 8000
)

Arguments

variant_df

A data frame containing the variant information.

variant_id_col

The name of the column in variant_df that contains the variant IDs.

effect_allele_col

The name of the column in variant_df that contains the effect alleles.

plink_bin

The path to the plink2 binary executable.

bfile

The prefix of the plink binary files (without the file extension).

threads

The number of threads to use for parallel processing (default: 1).

memory

The amount of memory (in MB) to allocate for plink2 (default: 8000).

Value

A long-format tibble::tibble() containing the extracted genotypes for each PMBB participant.

Examples

if (FALSE) {
pmbb_extract_genotypes(
  variant_df = ldlr_variants,
  variant_id_col = ID,
  effect_allele_col = Alt,
  plink_bin = "/project/voltron/Applications/PLINK/plink2_linux_avx2_20230607/plink2",
  bfile = "/project/PMBB/PMBB-Release-2020-2.0/Exome/pVCF/all_variants/PMBB-Release-2020-2.0_genetic_exome_GL"
)
}