Run a PheWAS and generate a report

This function is a wrapper around the run_pmbb_phewas function that additionally generates an HTML report of the results.

run_pmbb_phewas_report(
  gene,
  annotation_file,
  gene_col,
  masks,
  mask_operator,
  variant_id_col,
  effect_allele_col,
  plink_bin,
  bfile,
  phecode_file,
  covariate_files,
  populations,
  covariate_population_col,
  covariate_cols,
  mask_output_path = NULL,
  phewas_output_path = NULL,
  report_output_path = NULL,
  ...
)

Arguments

gene

The name of the gene to filter variants for.

annotation_file

The path to the variant annotation file.

gene_col

The name of the column in the annotation file to filter the gene on.

masks

A named list of lists containing filter criteria for each mask. Each element of the outer list should be named after the desired output name for the filtered data frame (eg. name of the variant mask). Each inner list should contain the filter criteria for a specific set of columns. For example: list(plof_0.001 = list(ExonicFunc.ensGene = "== 'stopgain'", gnomAD_exome_ALL = "< 0.001"), common_0.01 = list(gnomAD_exome_ALL = "> 0.01"))

mask_operator

A named list specifying the operation to perform for each mask. Valid operations are "burden" (default) and "single". The names in the mask_operator list should correspond to the names in the masks list. For example: list(plof_0.001 = "burden", common_0.01 = "single")

variant_id_col

The name of the column in the annotation file that contains the variant IDs.

effect_allele_col

The name of the column in the annotation file that contains the effect alleles.

plink_bin

The path to the plink2 binary executable.

bfile

The prefix of the plink binary files (without the file extension).

phecode_file

Path to PMBB phenotype file or a dataframe containing phenotype data used to run PheWAS

covariate_files

Path to PMBB covariate file(s) or a dataframe containing covariate data used to run PheWAS. If multiple files are provided, they will be joined by PMBB_ID, and values for duplicated columns will be preserved from the from the file that is specified earlier.

populations

A character vector of populations to run PheWAS on. Default is c("ALL")

covariate_population_col

Column containing population labels in the covariate file or dataframe, required if populations is not ALL

covariate_cols

Vector of columns in the covariate file or dataframe that should be used as covariates in the PheWAS

mask_output_path

(optional) Path to save the genotype output for each mask. If NULL the genotype output will not be saved (default: NULL)

phewas_output_path

(optional) Path to save the PheWAS output dataframe. If NULL the PheWAS output will not be saved separately from the report (default: NULL)

report_output_path

(optional) Path to save the HTML report. If NULL, the report will be saved in the working directory. (default: NULL)

...

Additional arguments passed to PheWAS::phewas()

Value

A named list containing the paths to the output files

Examples

if (FALSE) {
run_pmbb_phewas_report(
  gene = c("LDLR"),
  annotation_file = "/project/PMBB/PMBB-Release-2020-2.0/Exome/Variant_annotations/PMBB-Release-2020-2.0_genetic_exome_variant-annotation-counts.txt",
  gene_col = "Gene.refGene",
  masks = list(
    plof_0.01 = list(ExonicFunc.ensGene = "%in% c('stopgain', 'stoploss', 'frameshift substitution')", gnomAD_exome_ALL = "< 0.01"),
    missense_0.01 = list(ExonicFunc.ensGene = "%in% c('nonsynonymous SNV')", gnomAD_exome_ALL = "< 0.01")
  ),
  mask_operator = list(
    plof_0.01 = "burden",
    missense_0.01 = "burden"
  ),
  variant_id_col = ID,
  effect_allele_col = Alt,
  populations = c("ALL"),
  covariate_population_col = "Class",
  covariate_cols = c(Age = Age_at_Enrollment, Sex = Gen_Sex, dplyr::starts_with("Genotype_PC")),
  mask_output_path = "LDLR_mask_results.rds",  # Optional: Specify the output path for the mask results file
  phewas_output_path = "LDLR_phewas_results.rds",  # Optional: Specify the output path for the PheWAS results file,
  phecode_file = "/project/PMBB/PMBB-Release-2020-2.0/Phenotype/2.3/PMBB-Release-2020-2.3_phenotype_PheCode-matrix.txt",
  covariate_files = c("/project/PMBB/PMBB-Release-2020-2.0/Phenotype/2.3/PMBB-Release-2020-2.3_covariates.txt", "/project/PMBB/PMBB-Release-2020-2.0/Phenotype/2.1/PMBB-Release-2020-2.1_phenotype_covariates.txt"),
  bfile = "/project/PMBB/PMBB-Release-2020-2.0/Exome/pVCF/all_variants/PMBB-Release-2020-2.0_genetic_exome_GL",
  plink_bin = "/project/voltron/Applications/PLINK/plink2_linux_avx2_20230607/plink2",
  cores = 16
)
}