Run a PheWAS on PMBB data

This function runs a PheWAS on PMBB data, using genotype information and paths to PMBB files containing phenotype and covariate data. The function ideally accepts genotype information from the pmbb_extract_genotype_masks function, but can accept any genotype information of the format list(mask_name = list(genotypes = data.frame(PMBB_ID = character(), genotype = character())))

Usage

run_pmbb_phewas(
  mask_genotypes_list,
  phenotypes,
  covariates,
  populations = c("ALL"),
  covariate_cols,
  covariate_population_col = NULL,
  phenotype_case_count = 2,
  ...
)

Arguments

mask_genotypes_list

A named list, where each element contains a named element called genotypes which contains PMBB_ID and genotype columns.

phenotypes

Path to PMBB phenotype file or a dataframe containing phenotype data used to run PheWAS

covariates

Path to PMBB covariate file(s) or a dataframe containing covariate data used to run PheWAS. If multiple files are provided, they will be joined by PMBB_ID, and values for duplicated columns will be preserved from the from the file that is specified earlier.

populations

A character vector of populations to run PheWAS on. Default is c("ALL")

covariate_cols

Vector of columns in the covariate file or dataframe that should be used as covariates in the PheWAS

covariate_population_col

Column containing population labels in the covariate file or dataframe, required if populations is not ALL

phenotype_case_count

The minimum number of codes required to be considered a case (Default: 2)

...

Additional arguments passed to PheWAS::phewas()

Value

A tibble::tibble() containing the results of the PheWAS for each population and mask

Examples

if (FALSE) {
# Run a PheWAS
cftr_phewas_res <- pmbbPheWASr::run_pmbb_phewas(
  mask_genotypes_list = cftr_mask_res,
  populations = c("ALL", "EUR", "AFR"),
  covariate_population_col = "Class",
  covariate_cols = c(Age = Age_at_Enrollment, Sex = Gen_Sex, starts_with("Genotype_PC")),
  phenotypes = left_join(pmbb_phecodes_df, pmbb_labs_df),
  covariates = pmbb_covariates_df
)
}